%% Documentation of file 'tests/src/baseline_parameters.csv'

\name{baseline_parameters}
\docType{data}
\alias{baseline_parameters}
\title{Demographic Parameters}
\description{
  The demographic parameters are based from a number of different
  sources.

  For a detailed explanation of the available parameters (including
  sources and references), please read the
  \href{https://github.com/BDI-pathogens/OpenABM-Covid19/blob/master/documentation/parameters/parameter_dictionary.md}{Online Documentation}.
}
\usage{
data('baseline',package='OpenABMCovid19')
baseline_parameters
}
\references{
%% example
  G. E. P. Box and G. M. Jenkins (1976):
  \emph{Time Series Analysis, Forecasting and Control},
  Holden-Day, San Francisco, p.\sspace{}537.

  Joel Mossong, Niel Hens, Mark Jit, Philippe Beutels, Kari Auranen,
  Rafael Mikolajczyk, Marco Massari, Stefania Salmaso, Gianpaolo Scalia
  Tomba, Jacco Wallinga, et al. (2008): \emph{Social contacts and mixing
  patterns relevant to the spread of infectious diseases}, PLoSmedicine,
  5(3)

  Luca Ferretti, Chris Wymant, Michelle Kendall, Lele Zhao, Anel Nurtay,
  Lucie Abeler-Dorner, Michael Parker, David Bonsall, and Christophe
  Fraser (2020): \emph{Quantifying SARS-CoV-2 transmission suggests
  epidemic control  with  digital  contact tracing}, Science

  Qifang Bi, Yongsheng Wu, Shujiang Mei, Chenfei Ye, Xuan Zou, Zhen
  Zhang, Xiaojian Liu, Lan Wei, Shaun A Truelove, Tong Zhang, Wei Gao,
  Cong Cheng, Xiujuan Tang,Xiaoliang Wu, Yu Wu, Binbin Sun, Suli Huang,
  Yu Sun, Juncen Zhang, Ting Ma, JustinLessler,  and Teijian Feng
  (2020): \emph{Epidemiology and transmission of COVID-19 in
    ShenzhenChina:  Analysis of 391 cases and 1,286 of their close
    contacts}, medRxiv

  Lei Luo, Dan Liu, Xin-long Liao, Xian-bo Wu, Qin-long Jing, Jia-zhen
  Zheng, Fang-huaLiu, Shi-gui Yang, Bi Bi, Zhi-hao Li, Jian-ping Liu,
  Wei-qi Song, Wei Zhu,  Zheng-heWang, Xi-ru Zhang, Pei-liang Chen,
  Hua-min Liu, Xin Cheng, Miao-chun Cai, Qing-meiHuang, Pei Yang,
  Xin-fen Yang, Zhi-gang Huang, Jin-ling Tang, Yu Ma, and Chen Mao
  (2020): \emph{Modes of contact and risk of transmission in covid-19
  among close contacts}, medRxiv

  Yang Yang, Qingbin Lu, Mingjin Liu, Yixing Wang, Anran Zhang, Neda
  Jalali, NatalieDean, Ira Longini, M. Elizabeth Halloran, Bo Xu, Xiaoai
  Zhang, Liping Wang, Wei Liu, and Liqun Fang (2020):
  \emph{Epidemiological and clinical features of the 2019 novel
  coronavirus outbreak in China}. medRxiv.

  Juanjuan Zhang, Maria Litvinova, Yuxia Liang, Yan Wang, Wei Wang,
  Shanlu Zhao,Qianhui Wu, Stefano Merler, Cecile Viboud, Alessandro
  Vespignani, Marco Ajelli, and Hongjie Yu (2020): \emph{Age profile of
  susceptibility, mixing, and social distancing shape the dynamics of
  the novel coronavirus disease 2019 outbreak in china}, medRxiv

%% TODO(olegat) long citations needed, brief citations used for now.
  McAloon et al.

  Pellis et al, (2020)

  Personal communication with SPI-M; data soon to be published

  UK Office of National Statistics (2011):\emph{CT1089_2011 Census
  COVID-19}

  Calibration of Riccardo et al. 2020 & Spanish Serology Survey &
  Ferguson et al. (2020)

  Ferguson et al, (2020)

  Lu et al. 2020, Dong et al. (2020)

  Woelfel et al. (2020)

  OFCOM 3-5 year olds

  OFCOM 5-15 year olds

  OFCOM 16-55 year olds

  OFCOM 55+ year olds

  NACCHO Position Statement (2020)

  \url{https://www.gwhwi.org/estimator-613404.html}
}
