% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{create.params}
\alias{create.params}
\title{Parameters initializatoin wrapper}
\usage{
create.params(household, params)
}
\arguments{
\item{household}{Household demographics. Must be a data frame or a file path.
Required.}

\item{params}{Input parameters. Must be a data frame or a file path.
Required.}
}
\value{
An \code{\link{Parameters}} R6 instance.
}
\description{
Wrapper for creating \code{\link{Parameters}} instances. For legacy reasons
(to maintain feature-parity with the Python class interface), the constructor
R6 class takes arguments as file paths rather than data frames, this is not
also convenient when working in R. This wrapper function creates model
parameters using data frame arguments by default.

For a detailed explanation of the available parameters (including sources
and references), please read the
\href{https://github.com/BDI-pathogens/OpenABM-Covid19/blob/master/documentation/parameters/parameter_dictionary.md}{Online Documentation}.
}
\examples{
# Load data from baseline.rda
data('baseline', package='OpenABMCovid19')

# Create `Parameters` instance
params <- OpenABMCovid19::create.params(
  household = baseline_household_demographics,
  params    = baseline_parameters
)

# Edit params: method 1 (recommended, includes error-checking)
params$set_param('rng_seed', 1234)
params$set_param('n_total', 250000)

# Edit params: method 2 (not recommended, use with caution)
params$c_params$rng_seed = 1234
params$c_params$n_total = 250000

}
\seealso{
\code{\link{baseline_household_demographics}}

\code{\link{baseline_parameters}}
}
