%% Documentation of file 'tests/src/hospital_baseline_parameters.csv'

\name{hospital_baseline_parameters}
\docType{data}
\alias{hospital_baseline_parameters}
\title{Hospital Parameters}
\description{
  For information about this data please read the
  \href{https://github.com/BDI-pathogens/OpenABM-Covid19/blob/master/documentation/hospital_model.pdf}{Online Documentation}.
}
\usage{
data('baseline',package='OpenABMCovid19')
hospital_baseline_parameters
}
\source{
  \url{https://github.com/BDI-pathogens/OpenABM-Covid19/blob/master/documentation/hospital_model.pdf}
}
