% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model.get_param}
\alias{Model.get_param}
\title{Gets the value of a parameter (wrapper for \
code{\link{Parameters}$get_param(param)})}
\usage{
Model.get_param(model, param)
}
\arguments{
\item{param}{The name of the parameter}

\item{parameters}{A Parameters object}
}
\description{
Gets the value of a parameter (wrapper for \
code{\link{Parameters}$get_param(param)})
}
