% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model.get_transmissions}
\alias{Model.get_transmissions}
\title{Gets all the transmissions until now (wrapper for
\code{\link{Model}$get_transmissions()})}
\usage{
Model.get_transmissions(model)
}
\arguments{
\item{model}{The Model object (R6 Class)}
}
\value{
DataFrame
}
\description{
Gets all the transmissions until now (wrapper for
\code{\link{Model}$get_transmissions()})
}
