% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model.new}
\alias{Model.new}
\title{Creates a new OpenABM Model (wrapper for \code{\link{Model}$new()})}
\usage{
Model.new(params_object = NULL, params = NULL)
}
\arguments{
\item{params_object}{An object of type \code{\link{Parameters}} or NULL
(for default params). The constructor will lock the parameter values (ie.
\code{params_code} will become read-only).}

\item{params}{A named list of parameters fo override (default NULL for no
overrides)}
}
\value{
Model object (R6 Class), NULL on error or if out-of-memory.
}
\description{
Creates a new OpenABM \code{\link{Model}} instance from a
\code{\link{Parameters}} object and/or a list of parameters overrides.

Note that model instances are a memory/CPU extensive and this function
can take a few seconds to complete.
}
