% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model.one_time_step}
\alias{Model.one_time_step}
\title{Steps the model forward one time step (wrapper for \code{\link{Model}$one_time_step()})}
\usage{
Model.one_time_step(model)
}
\arguments{
\item{model}{The Model object (R6 Class)}
}
\value{
Null
}
\description{
Steps the model forward one time step (wrapper for \code{\link{Model}$one_time_step()})
}
