% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model.one_time_step_results}
\alias{Model.one_time_step_results}
\title{Gets the simulation results for the current time-step. (wrapper for
\code{\link{Model}$one_time_step_results()})}
\usage{
Model.one_time_step_results(model)
}
\arguments{
\item{model}{The Model object (R6 Class)}
}
\value{
Vector with names of variables
}
\description{
Gets the simulation results for the current time-step. (wrapper for
\code{\link{Model}$one_time_step_results()})
}
