% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model.results}
\alias{Model.results}
\title{Gets the simulation results for all time-steps run so far (wrapper for
\code{\link{Model}$results()})}
\usage{
Model.results(model)
}
\arguments{
\item{model}{The Model object (R6 Class)}
}
\value{
DataFrame
}
\description{
Gets the simulation results for all time-steps run so far (wrapper for
\code{\link{Model}$results()})
}
