% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model.run}
\alias{Model.run}
\title{Runs the simulation until the \code{end_time} specified in the parameters
(wrapper for  \code{\link{Model}run()})}
\usage{
Model.run(model, verbose = TRUE)
}
\arguments{
\item{model}{The Model object (R6 Class)}

\item{verbose}{Show progress of the calculation (default = TRUE)}
}
\value{
Null
}
\description{
Runs the simulation until the \code{end_time} specified in the parameters
(wrapper for  \code{\link{Model}run()})
}
