% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model.update_running_params}
\alias{Model.update_running_params}
\title{Update a parameter during a simulation}
\usage{
Model.update_running_params(model, param, value)
}
\arguments{
\item{model}{A Model object}

\item{param}{name of parameter. See \code{\link{SAFE_UPDATE_PARAMS}} for
allowed parameter names}

\item{value}{value of parameter}
}
\description{
A subset of parameters may be updated whilst the model is
evaluating these correspond to events. This function throws an error if
\code{param} isn't safe to update.
}
