% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{Parameters.get_param}
\alias{Parameters.get_param}
\title{Gets the value of a parameter (wrapper for
\code{\link{Parameters}$get_param(param)})}
\usage{
Parameters.get_param(parameters, param)
}
\arguments{
\item{parameters}{A Parameters object}

\item{param}{The name of the parameter}
}
\description{
Gets the value of a parameter (wrapper for
\code{\link{Parameters}$get_param(param)})
}
\seealso{
\href{https://github.com/BDI-pathogens/OpenABM-Covid19/blob/master/documentation/parameters/parameter_dictionary.md}{Online Documentation}.
}
