% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{Parameters.set_param}
\alias{Parameters.set_param}
\title{Sets the value of a parameter (wrapper for
\code{\link{Parameters}$set_param(param,value)})}
\usage{
Parameters.set_param(parameters, param, value)
}
\arguments{
\item{parameters}{A Parameters object}

\item{param}{The name of the parameter}

\item{value}{The new value}
}
\description{
Sets the value of a parameter (wrapper for
\code{\link{Parameters}$set_param(param,value)})
}
\seealso{
\href{https://github.com/BDI-pathogens/OpenABM-Covid19/blob/master/documentation/parameters/parameter_dictionary.md}{Online Documentation}.
}
