%% This file contains S4 methods are auto-generated by SWIG
%% in R/OpenABMCovid19.R
%%
%% `R CMD check` requires that all user-level methods have, at the
%% very least, an \alias in the documentation. This file contains those
%% \alias commands explains what they do and where they come from.
%%
\name{swig_methods}
\title{SWIG Generated S4 methods}

%% OpenABMCovid19 types:
\alias{$,_p_directory-method}
\alias{$,_p_doubleArray-method}
\alias{$,_p_edge-method}
\alias{$,_p_event-method}
\alias{$,_p_event_block-method}
\alias{$,_p_event_list-method}
\alias{$,_p_incomplete_gamma_p_params-method}
\alias{$,_p_individual-method}
\alias{$,_p_infection_event-method}
\alias{$,_p_intArray-method}
\alias{$,_p_interaction-method}
\alias{$,_p_interaction_block-method}
\alias{$,_p_longArray-method}
\alias{$,_p_model-method}
\alias{$,_p_network-method}
\alias{$,_p_parameters-method}
\alias{$,_p_shortArray-method}
\alias{$,_p_trace_token-method}
\alias{$,_p_trace_token_block-method}
\alias{$<-,_p_directory-method}
\alias{$<-,_p_edge-method}
\alias{$<-,_p_event-method}
\alias{$<-,_p_event_block-method}
\alias{$<-,_p_event_list-method}
\alias{$<-,_p_incomplete_gamma_p_params-method}
\alias{$<-,_p_individual-method}
\alias{$<-,_p_infection_event-method}
\alias{$<-,_p_interaction-method}
\alias{$<-,_p_interaction_block-method}
\alias{$<-,_p_model-method}
\alias{$<-,_p_network-method}
\alias{$<-,_p_parameters-method}
\alias{$<-,_p_trace_token-method}
\alias{$<-,_p_trace_token_block-method}
\alias{[[<-,_p_directory,character-method}
\alias{[[<-,_p_edge,character-method}
\alias{[[<-,_p_event,character-method}
\alias{[[<-,_p_event_block,character-method}
\alias{[[<-,_p_event_list,character-method}
\alias{[[<-,_p_incomplete_gamma_p_params,character-method}
\alias{[[<-,_p_individual,character-method}
\alias{[[<-,_p_infection_event,character-method}
\alias{[[<-,_p_interaction,character-method}
\alias{[[<-,_p_interaction_block,character-method}
\alias{[[<-,_p_model,character-method}
\alias{[[<-,_p_network,character-method}
\alias{[[<-,_p_parameters,character-method}
\alias{[[<-,_p_trace_token,character-method}
\alias{[[<-,_p_trace_token_block,character-method}

%% SWIG helper types
\alias{[,ExternalReference-method}
\alias{[<-,ExternalReference-method}
\alias{length,SWIGArray-method}

\description{
  Extractor and replace methods auto-generated by SWIG.
}

\section{Classes}{
  Class names prefixes with \code{_p_} are wrappers to their respective
  C type in the dynamic library. Example: \R class \code{_p_model} maps
  to the C struct \code{model}.

  Classes \code{ExternalReference} and \code{SWIGArray} are helper
  classes that are not part of the model framework.
}
